refect(interface):interface타입 정의를 따로 정리
