+++ src/server.ts
@@ -1,4 +1,3 @@
-
 import 'dotenv/config';
 import app from './app.js';
